% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_clustered.bar.R
\name{chem_clustered.bar}
\alias{chem_clustered.bar}
\title{Creates a clustered bar graph}
\usage{
chem_clustered.bar(data, xvar, yvar, fillvar, xlab = "Insert X-axis label",
  ylab = "Insert Y-axis label", n = 4, colors = hcl(h = seq(15, 375,
  length = n + 1), l = 65, c = 100)[1:n])
}
\arguments{
\item{data}{A data frame containing variables to be plotted}

\item{xvar}{The name of the x-variable}

\item{yvar}{The name of the y-variable}

\item{fillvar}{The name of the variable to color the bars by}

\item{xlab}{A string containing the x-axis label}

\item{ylab}{A string containing the y-axis label}

\item{n}{The number of unique entries in \code{fillvar}}

\item{colors}{A character vector containing the names of the
R colors to color the plot with}
}
\value{
A plot
}
\description{
Creates a clustered bar graph
}
