% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_dual.scatter.R
\name{chem_dual.scatter}
\alias{chem_dual.scatter}
\title{Creates a scatterplot with a common horizontal axis and two different
variables plotted on the vertical axis with regression lines}
\usage{
chem_dual.scatter(data, xvar, yvar1, yvar2, xlab = "Insert X-axis label",
  ylab = "Insert Y-axis label", intercept1 = NULL, intercept2 = NULL,
  size = 3, colors = c("red", "blue"))
}
\arguments{
\item{data}{A data frame containing variables to be plotted}

\item{xvar}{The name of the x-variable}

\item{yvar1}{The name of one of the y-variables}

\item{yvar2}{The name of the other y-variable}

\item{xlab}{A string containing the x-axis label}

\item{ylab}{A string containing the y-axis label}

\item{intercept1}{Set to NULL by default, but can be changed to 0 to force regression
line to go through origin for yvar1}

\item{intercept2}{Set to NULL by default, but can be changed to 0 to force regression
line to go through origin for yvar2}

\item{size}{A numeric specifying the size of the points}

\item{colors}{A character vector specifying colors of two y-variables and each line}
}
\value{
A plot
}
\description{
Creates a scatterplot with a common horizontal axis and two different
variables plotted on the vertical axis with regression lines
}
