% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_scatter.R
\name{chem_scatter}
\alias{chem_scatter}
\title{Creates a scatterplot with regression line}
\usage{
chem_scatter(data, xvar, yvar, xlab = "Insert X-axis label",
  ylab = "Insert Y-axis label", intercept = NULL)
}
\arguments{
\item{data}{A data frame containing variables to be plotted}

\item{xvar}{The name of the x-variable}

\item{yvar}{The name of the y-variable}

\item{xlab}{A string containing the x-axis label}

\item{ylab}{A string containing the y-axis label}

\item{intercept}{Set to NULL by default, but can be changed to 0 to
force regression line to go through origin}
}
\value{
A plot
}
\description{
Creates a scatterplot with regression line
}
\examples{
\dontrun{
 chem_scatter(iris, Sepal.Width, Sepal.Length)
}
}

