% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_lab.R
\name{chem_lab}
\alias{chem_lab}
\title{Creates an R Markdown PDF document for the Beers Law Lab}
\usage{
chem_lab(fig_height = 2.5, fig_width = 5)
}
\arguments{
\item{fig_height}{A positive numeric giving the default height of images}

\item{fig_width}{A positive numeric giving the default width of images}
}
\value{
A modified \code{pdf_document} based on slightly modified knitr
LaTeX default file
}
\description{
This is a function called in output in the YAML of the driver Rmd file to
specify using a tweaked version of the default LaTeX pandoc template.
}
\examples{
\dontrun{
 output:
   chemistr::chem_lab
}
}

