% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_table.R
\name{chem_table}
\alias{chem_table}
\title{Creates a nice table for data frames}
\usage{
chem_table(data, caption)
}
\arguments{
\item{data}{A data frame containing variables to be plotted}

\item{x}{A character string of column names}
}
\value{
A string of bolded column names
}
\description{
Creates a nice table for data frames
}
\examples{
\dontrun{
 chem_table(head(iris), caption = "First six rows of iris data")
}
}

